@@ -6,13 +6,13 @@ dependencies {
 android {
     //compileSdkVersion Integer.parseInt(project.ANDROID_BUILD_SDK_VERSION)
     //buildToolsVersion project.ANDROID_BUILD_TOOLS_VERSION
-    compileSdkVersion 25
-    buildToolsVersion "24.0.3"
+    compileSdkVersion 29
+    buildToolsVersion '29.0.2'
     defaultConfig {
         //minSdkVersion Integer.parseInt(project.ANDROID_BUILD_MIN_SDK_VERSION)
         //targetSdkVersion Integer.parseInt(project.ANDROID_BUILD_TARGET_SDK_VERSION)
         minSdkVersion 15
-        targetSdkVersion 25
+        targetSdkVersion 29
     }
     sourceSets {
         main {
