@@ -1,6 +1,6 @@
 package com.example.newsAPP.activity;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 
 /**
