@@ -3,8 +3,8 @@ package com.example.newsAPP.fragment;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.Nullable;
+import androidx.appcompat.widget.Toolbar;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,11 +13,9 @@ import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.example.newsAPP.R;
 import com.example.newsAPP.activity.AboutActivity;
-import com.example.newsAPP.activity.FeedbackActivity;
 import com.example.newsAPP.activity.FriendListActivity;
 import com.example.newsAPP.activity.LoginActivity;
 import com.example.newsAPP.activity.SettingActivity;
