@@ -2,7 +2,7 @@ package com.example.newsAPP.widget;
 
 import android.content.Context;
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.example.newsAPP.R;
