@@ -1,6 +1,6 @@
 package com.example.channelmanager.base;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.MotionEvent;
 
 import com.example.channelmanager.adapter.ChannelAdapter;
